requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
Pillow==10.1.0
exifread==3.0.0
requests==2.31.0
openai==2.6.1
gunicorn==21.2.0
flask-cors==4.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
c2pa-python==0.4.0
